+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Disable warnings during RunHandler initialization to avoid showing warnings with respect to previously loaded runs.
 - Remove option to use docker (not yet supported).
 - Add SMAC3v2 multi-fidelity example runs.
+- Add more useful SMAC3v2 multi-objective example run.
 
 # Version 1.2.1